+++ internal/dirlock/lock_unix.go
@@ -1,0 +1,10 @@
+//go:build !windows
+
+package dirlock
+
+import (
+	"golang.org/x/sys/unix"
+	"os"
+)
+
+func lock(f *os.File) error { return unix.Flock(int(f.Fd()), unix.LOCK_EX|unix.LOCK_NB) }